tambah konfirmasi dialog

+++ src/components/PokemonCards.js
@@ -6,14 +6,17 @@ export function PokemonCards({ pokemon }) {
 
     function release(nickname, e) {
         e.preventDefault()
+        const remove = window.confirm(`Kamu yakin akan melepas ${nickname}?`)
         const myPokemon = JSON.parse(localStorage.getItem('mypokemon'))
-        for(let i=0; i < myPokemon.length; i++) {
-            if (myPokemon[i].nickname === nickname) {
-                myPokemon.splice(i,1)
+        if (remove) {
+            for(let i=0; i < myPokemon.length; i++) {
+                if (myPokemon[i].nickname === nickname) {
+                    myPokemon.splice(i,1)
+                }
             }
+            localStorage.setItem('mypokemon', JSON.stringify(myPokemon))
+            window.location.reload()
         }
-        localStorage.setItem('mypokemon', JSON.stringify(myPokemon))
-        window.location.reload()
     }
     
     return (
